Release nxtvepg-0.4.pre0 (2000-12-26)
Changes between 0.3.4 and 0.4.pre0 (as per file CHANGES):
- added possibility to merge several provider databases into one
- cyclic acquisition between several databases
- added TV input configuration popup
- added Help button to all popups
  order of Help, Abort, Ok identical in all popups
- port to M$ Windows (in new module epgvbi/btdrv4win.c)

NOTE: The WIN32 build requires header files windrvr.h and windrvr_usb.h
which were not included in source code releases and are also not in this
repository for copyright reasons. See file README.

See file CHANGES.RCS for a detailed log of revisions of all files.