Release nxtvepg-0.4.pre1 (2001-Jan-09)
Changes between 0.4.pre0 and 0.4.pre1 (as per file CHANGES):
- presentation of descriptions of merged databases improved
  by adding a separator image (horizontal line) between the texts.
- added support for Net/Free/OpenBSD Unix in btdrv4linux by Mario Kemper
- described Windows installation in more depth and length in the README file
- compatibility with WinDriver 4.00
- updated registry hack for Windows Bt8x8 driver 4.31 and later
- command line can take database path as last argument
  This is intended for Windows, to be able to drag and drop a database file
  onto the executable, but can generally be used to set -dbdir and -provider
- added warning messages for database reload errors or missing frequencies

NOTE: The WIN32 build requires header files windrvr.h and windrvr_usb.h
which were not included in source code releases and are also not in this
repository for copyright reasons. See file README.

See file CHANGES.RCS for a detailed log of revisions of all files.