Release nxtvepg-0.4.pre2 (2001-Jan-21)
Changes between 0.4.pre1 and 0.4.pre2 (as per file CHANGES):
- Added status line below browser window for db coverage and acq status
- Statistics windows now available separately for browser and acq databases.
  Added pie chart in addition to history diagram.
- Acquisition statistics are now updated even while no EPG reception, to
  allow analysis of reception trouble. Added output of VPS/PDC CNI.
- NetBSD: patch from Mario to fix possible hangup of VBI slave after an
  external /dev/video application is terminated.
- Windows Bt8x8 driver: added PLL init for 35MHz; Improved driver start on NT;
  Fixed blue screen during shutdown and crashes: exceptions and the shutdown
  message are now catched and the driver stopped before the program exits;
  Disabled registry hack for legal reasons.

NOTE: The WIN32 build requires header files windrvr.h and windrvr_usb.h
which were not included in source code releases and are also not in this
repository for copyright reasons. See file README.

See file CHANGES.RCS for a detailed log of revisions of all files.