TV-simulator: Added option -fakechantab
Added an option "-fakechantab" to the TV-application simulator, so that it
can also be used for demo and testing purpose on systems without a TV card
and thus without installed TV application. Documented the new option in
the manual.

Also updated the "winshm" protocol version equivalently to XICCM version,
as the EPG info format has changed recently.